+++ samples/ContactSync/ContactSync.OldApp/ContactSync.OldApp.Database/Migrations/Setup.post.database.create.sql
@@ -1,16 +1,4 @@
-﻿/*
- Pre-Deployment Script Template							
---------------------------------------------------------------------------------------
- This file contains SQL statements that will be executed before the build script.	
- Use SQLCMD syntax to include a file in the pre-deployment script.			
- Example:      :r .\myfile.sql								
- Use SQLCMD syntax to reference a variable in the pre-deployment script.		
- Example:      :setvar TableName MyTable							
-               SELECT * FROM [$(TableName)]					
---------------------------------------------------------------------------------------
-*/
-
--- Enable CDC for database
+﻿-- Enable CDC for database
 
 IF (SELECT TOP 1 is_cdc_enabled FROM sys.databases WHERE [name] = DB_NAME()) = 0
 BEGIN